+++ config/initializers/container.rb
@@ -49,6 +49,7 @@ module Charkeeper
     register('commands.characters_context.cosmere.create') { CharactersContext::Cosmere::CreateCommand.new }
     register('commands.characters_context.cosmere.update') { CharactersContext::Cosmere::UpdateCommand.new }
     register('commands.characters_context.cosmere.rest.perform') { CharactersContext::Cosmere::Rest::PerformCommand.new }
+    register('commands.characters_context.cosmere.feats.add') { CharactersContext::Cosmere::Feats::AddCommand.new }
 
     register('commands.characters_context.dc20.create') { CharactersContext::Dc20::CreateCommand.new }
     register('commands.characters_context.dc20.update') { CharactersContext::Dc20::UpdateCommand.new }