+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - leveling for Cosmere characters
 - selecting heroic path for Cosmere characters
 - selecting expertises for Cosmere characters
+- selecting talents for Cosmere characters
 
 ## [0.4.29] - 2026-04-21
 ### Added